Replace references to "General DigiMusic".
The official name for the GDM format in all Bells, Whistles, and
Sound Boards documentation and source is "General Digital Music".

I'm not sure where the term "General DigiMusic" came from, but it
does not appear in BWSB at all. It was used by Alexis Janson in the
MegaZeux documentation in the '90s, so it might have been invented
by her, but I don't have any evidence of this.

This term probably found its way here because Kevin Vance, who
contributed the initial GDM loader, was a member of the MZX community.

There was also a typo ("Munsic") in the manpage that I fixed.